Fixing guessed key output. Only in Guess Mode
